+++ protocols/identify/src/handler.rs
@@ -412,9 +412,6 @@ impl ConnectionHandler for Handler {
                 ));
                 self.trigger_next_identify.reset(self.interval);
             }
-            ConnectionEvent::AddressChange(_)
-            | ConnectionEvent::ListenUpgradeError(_)
-            | ConnectionEvent::RemoteProtocolsChange(_) => {}
             ConnectionEvent::LocalProtocolsChange(change) => {
                 let before = log::log_enabled!(Level::Debug)
                     .then(|| self.local_protocols_to_string())
@@ -439,6 +436,7 @@ impl ConnectionHandler for Handler {
                         });
                 }
             }
+            _ => {}
         }
     }
 }