+++ protocols/kad/src/handler.rs
@@ -792,9 +792,6 @@ impl ConnectionHandler for Handler {
             ConnectionEvent::DialUpgradeError(dial_upgrade_error) => {
                 self.on_dial_upgrade_error(dial_upgrade_error)
             }
-            ConnectionEvent::AddressChange(_)
-            | ConnectionEvent::ListenUpgradeError(_)
-            | ConnectionEvent::LocalProtocolsChange(_) => {}
             ConnectionEvent::RemoteProtocolsChange(change) => {
                 let dirty = self.remote_supported_protocols.on_protocols_change(change);
 
@@ -828,6 +825,7 @@ impl ConnectionHandler for Handler {
                     }
                 }
             }
+            _ => {}
         }
     }
 }