+++ swarm/src/behaviour.rs
@@ -237,6 +237,7 @@ pub trait PollParameters {
 ///
 /// [`Swarm`]: super::Swarm
 #[derive(Debug)]
+#[non_exhaustive]
 pub enum ToSwarm<TOutEvent, TInEvent> {
     /// Instructs the `Swarm` to return an event when it is being polled.
     GenerateEvent(TOutEvent),
@@ -409,6 +410,7 @@ pub enum CloseConnection {
 /// Enumeration with the list of the possible events
 /// to pass to [`on_swarm_event`](NetworkBehaviour::on_swarm_event).
 #[derive(Debug)]
+#[non_exhaustive]
 pub enum FromSwarm<'a, Handler> {
     /// Informs the behaviour about a newly established connection to a peer.
     ConnectionEstablished(ConnectionEstablished<'a>),